Update RegisterConsulService to wait and poll for service registrations
Most of our e2e tests with call time.Sleep() after calling
RegisterConsulService. Rather than wait an arbitrary amount of time, add polling
in RegisterConsulService to wait and poll GET '/agent/service/:id' until the
service is registered. RegisterConsulService() can be set with a liberal wait
timeout.

This will be consumed in a later commit alongside WaitForEvent()

Changes:
 - Move existing RegisterConsulService from utils.go => consul.go
 - Add new function to check if a service is registered
 - Update RegisterConsulService() to register service and then poll to see if
 service is registered
